Remove `DOCUMENTER_KEY` for regular documentation deployment
This secret is not actually necessary for deployment to work in this
case. The `GITHUB_TOKEN` can also be used to push to the branch. Now
that `DOCUMENTER_KEY` actually has a value and is set to an non-base64
encoded SSH key the deployment errors[^1].

Partially reverts 048a234c21483ec1297d31978f4e52812b25a522.

[^1]: https://github.com/JuliaComputing/Multibody.jl/actions/runs/4697514324/jobs/8328667586